requirements.txt
# Vietnamese Sign Language Detection - Core Requirements
opencv-python>=4.5.0
tensorflow>=2.8.0
mediapipe>=0.8.0
numpy>=1.21.0
scikit-learn>=1.0.0
pyyaml>=6.0
pillow>=8.0.0
matplotlib>=3.5.0
PyQt5>=5.15.0
gtts>=2.2.0
pygame>=2.1.0
